pyproject.toml
[tool.poetry]
name = "metagnn"
version = "0.1.0"
description = "A GNN-based framework for metagenomics."
authors = ["William O'Brien <wob@cs.ucla.edu>", "Seungmo Lee <seungmo6@g.ucla.edu>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/wgobrien/metagnn"
repository = "https://github.com/wgobrien/metagnn"
keywords = ["metagenomics", "GNN", "bioinformatics"]
packages = [
    { include = "metagnn", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
pyro-ppl = "^1.9.1"
torch = "^2.0.0"
torch-geometric = "^2.1.0"
pandas = "^1.5.0"
seaborn = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^23.0"

[tool.poetry.scripts]
metagnn = "metagnn.metagnn:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"